rename: 스토리북을 위한 디렉토리 구조 변경
